requirements.txt
# Pastor Noticias - Dependencias del proyecto

# Framework principal
streamlit>=1.28.0
streamlit-autorefresh>=0.0.1
selenium>=4.11.0
webdriver-manager>=4.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Base de datos
mysql-connector-python>=8.1.0
pymysql>=1.1.0

# Análisis de datos
pandas>=2.0.0
numpy>=1.24.0

# Visualizaciones
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Procesamiento de texto
wordcloud>=1.9.0
nltk>=3.8.0

# Utilidades
schedule>=1.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Desarrollo y testing (opcional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Extras para deployment
gunicorn>=21.2.0
uvicorn>=0.23.0